[Feat] 사용자 섹션 분리
